https://raw.githubusercontent.com/arib168/data/main/50_Startups.csv
Perform regression analysis to predict the profit of each the 50 startups
mentioned in the dataset by taking all the input parameters (multi linear
regression).
Use the dataset given below. The dataset consists of 5 columns where the ‘state’
column stores the categorical data points, and the rest are numerical features. Right-click and click on save as (50_starups.csv) to download the dataset